Allowing primary displayed item to dynamically change

+++ davega_ili9341_screen.h
@@ -27,8 +27,8 @@
 #include "davega_screen.h"
 
 typedef struct {
-    uint8_t x;
-    uint8_t y;
+    uint16_t x;
+    uint16_t y;
 } Point;
 
 typedef struct {
